TecnicosEntity.kt ha sido movido a su respectiva carpeta
